moved color_picker and system statistcs functions to model_utils
