+++ app.py
@@ -4,31 +4,13 @@ import torch
 from utils.hubconf import custom
 import numpy as np
 import tempfile
-from PIL import ImageColor
 import time
 from collections import Counter
 import json
-import psutil
-import subprocess
 import pandas as pd
-from model_utils import get_yolo
+from model_utils import get_yolo, color_picker_fn, get_system_stat
 from ultralytics import YOLO
 
-
-def get_gpu_memory():
-    result = subprocess.check_output(
-        [
-            'nvidia-smi', '--query-gpu=memory.used',
-            '--format=csv,nounits,noheader'
-        ], encoding='utf-8')
-    gpu_memory = [int(x) for x in result.strip().split('\n')]
-    return gpu_memory[0]
-
-def color_picker_fn(classname, key):
-    color_picke = st.sidebar.color_picker(f'{classname}:', '#ff0003', key=key)
-    color_rgb_list = list(ImageColor.getcolor(str(color_picke), "RGB"))
-    color = [color_rgb_list[2], color_rgb_list[1], color_rgb_list[0]]
-    return color
 
 p_time = 0
 
@@ -100,7 +82,7 @@ if not model_type == 'YOLO Model':
             upload_img_file = st.sidebar.file_uploader(
                 'Upload Image', type=['jpg', 'jpeg', 'png'])
             if upload_img_file is not None:
-                pred = st.checkbox('Predict Using YOLOv7')
+                pred = st.checkbox(f'Predict Using {model_type}')
                 file_bytes = np.asarray(
                     bytearray(upload_img_file.read()), dtype=np.uint8)
                 img = cv2.imdecode(file_bytes, 1)
@@ -161,44 +143,7 @@ if not model_type == 'YOLO Model':
                         df_fq = pd.DataFrame(class_fq.items(), columns=['Class', 'Number'])
                         
                         # Updating Inference results
-                        with stframe1.container():
-                            st.markdown("<h2>Inference Statistics</h2>", unsafe_allow_html=True)
-                            if round(fps, 4)>1:
-                                st.markdown(f"<h4 style='color:green;'>Frame Rate: {round(fps, 4)}</h4>", unsafe_allow_html=True)
-                            else:
-                                st.markdown(f"<h4 style='color:red;'>Frame Rate: {round(fps, 4)}</h4>", unsafe_allow_html=True)
-                        
-                        with stframe2.container():
-                            st.markdown("<h3>Detected objects in curret Frame</h3>", unsafe_allow_html=True)
-                            st.dataframe(df_fq, use_container_width=True)
-
-                        with stframe3.container():
-                            st.markdown("<h2>System Statistics</h2>", unsafe_allow_html=True)
-                            js1, js2, js3 = st.columns(3)                       
-
-                            # Updating System stats
-                            with js1:
-                                st.markdown("<h4>Memory usage</h4>", unsafe_allow_html=True)
-                                mem_use = psutil.virtual_memory()[2]
-                                if mem_use > 50:
-                                    js1_text = st.markdown(f"<h5 style='color:red;'>{mem_use}%</h5>", unsafe_allow_html=True)
-                                else:
-                                    js1_text = st.markdown(f"<h5 style='color:green;'>{mem_use}%</h5>", unsafe_allow_html=True)
-
-                            with js2:
-                                st.markdown("<h4>CPU Usage</h4>", unsafe_allow_html=True)
-                                cpu_use = psutil.cpu_percent()
-                                if mem_use > 50:
-                                    js2_text = st.markdown(f"<h5 style='color:red;'>{cpu_use}%</h5>", unsafe_allow_html=True)
-                                else:
-                                    js2_text = st.markdown(f"<h5 style='color:green;'>{cpu_use}%</h5>", unsafe_allow_html=True)
-
-                            with js3:
-                                st.markdown("<h4>GPU Memory Usage</h4>", unsafe_allow_html=True)  
-                                try:
-                                    js3_text = st.markdown(f'<h5>{get_gpu_memory()} MB</h5>', unsafe_allow_html=True)
-                                except:
-                                    js3_text = st.markdown('<h5>NA</h5>', unsafe_allow_html=True)
+                        get_system_stat(stframe1, stframe2, stframe3, fps, df_fq)
 
 
         # Web-cam
@@ -237,44 +182,7 @@ if not model_type == 'YOLO Model':
                         df_fq = pd.DataFrame(class_fq.items(), columns=['Class', 'Number'])
                         
                         # Updating Inference results
-                        with stframe1.container():
-                            st.markdown("<h2>Inference Statistics</h2>", unsafe_allow_html=True)
-                            if round(fps, 4)>1:
-                                st.markdown(f"<h4 style='color:green;'>Frame Rate: {round(fps, 4)}</h4>", unsafe_allow_html=True)
-                            else:
-                                st.markdown(f"<h4 style='color:red;'>Frame Rate: {round(fps, 4)}</h4>", unsafe_allow_html=True)
-                        
-                        with stframe2.container():
-                            st.markdown("<h3>Detected objects in curret Frame</h3>", unsafe_allow_html=True)
-                            st.dataframe(df_fq, use_container_width=True)
-
-                        with stframe3.container():
-                            st.markdown("<h2>System Statistics</h2>", unsafe_allow_html=True)
-                            js1, js2, js3 = st.columns(3)                       
-
-                            # Updating System stats
-                            with js1:
-                                st.markdown("<h4>Memory usage</h4>", unsafe_allow_html=True)
-                                mem_use = psutil.virtual_memory()[2]
-                                if mem_use > 50:
-                                    js1_text = st.markdown(f"<h5 style='color:red;'>{mem_use}%</h5>", unsafe_allow_html=True)
-                                else:
-                                    js1_text = st.markdown(f"<h5 style='color:green;'>{mem_use}%</h5>", unsafe_allow_html=True)
-
-                            with js2:
-                                st.markdown("<h4>CPU Usage</h4>", unsafe_allow_html=True)
-                                cpu_use = psutil.cpu_percent()
-                                if mem_use > 50:
-                                    js2_text = st.markdown(f"<h5 style='color:red;'>{cpu_use}%</h5>", unsafe_allow_html=True)
-                                else:
-                                    js2_text = st.markdown(f"<h5 style='color:green;'>{cpu_use}%</h5>", unsafe_allow_html=True)
-
-                            with js3:
-                                st.markdown("<h4>GPU Memory Usage</h4>", unsafe_allow_html=True)  
-                                try:
-                                    js3_text = st.markdown(f'<h5>{get_gpu_memory()} MB</h5>', unsafe_allow_html=True)
-                                except:
-                                    js3_text = st.markdown('<h5>NA</h5>', unsafe_allow_html=True)
+                        get_system_stat(stframe1, stframe2, stframe3, fps, df_fq)
 
 
         # RTSP
@@ -320,41 +228,4 @@ if not model_type == 'YOLO Model':
                     df_fq = pd.DataFrame(class_fq.items(), columns=['Class', 'Number'])
                     
                     # Updating Inference results
-                    with stframe1.container():
-                        st.markdown("<h2>Inference Statistics</h2>", unsafe_allow_html=True)
-                        if round(fps, 4)>1:
-                            st.markdown(f"<h4 style='color:green;'>Frame Rate: {round(fps, 4)}</h4>", unsafe_allow_html=True)
-                        else:
-                            st.markdown(f"<h4 style='color:red;'>Frame Rate: {round(fps, 4)}</h4>", unsafe_allow_html=True)
-                    
-                    with stframe2.container():
-                        st.markdown("<h3>Detected objects in curret Frame</h3>", unsafe_allow_html=True)
-                        st.dataframe(df_fq, use_container_width=True)
-
-                    with stframe3.container():
-                        st.markdown("<h2>System Statistics</h2>", unsafe_allow_html=True)
-                        js1, js2, js3 = st.columns(3)                       
-
-                        # Updating System stats
-                        with js1:
-                            st.markdown("<h4>Memory usage</h4>", unsafe_allow_html=True)
-                            mem_use = psutil.virtual_memory()[2]
-                            if mem_use > 50:
-                                js1_text = st.markdown(f"<h5 style='color:red;'>{mem_use}%</h5>", unsafe_allow_html=True)
-                            else:
-                                js1_text = st.markdown(f"<h5 style='color:green;'>{mem_use}%</h5>", unsafe_allow_html=True)
-
-                        with js2:
-                            st.markdown("<h4>CPU Usage</h4>", unsafe_allow_html=True)
-                            cpu_use = psutil.cpu_percent()
-                            if mem_use > 50:
-                                js2_text = st.markdown(f"<h5 style='color:red;'>{cpu_use}%</h5>", unsafe_allow_html=True)
-                            else:
-                                js2_text = st.markdown(f"<h5 style='color:green;'>{cpu_use}%</h5>", unsafe_allow_html=True)
-
-                        with js3:
-                            st.markdown("<h4>GPU Memory Usage</h4>", unsafe_allow_html=True)  
-                            try:
-                                js3_text = st.markdown(f'<h5>{get_gpu_memory()} MB</h5>', unsafe_allow_html=True)
-                            except:
-                                js3_text = st.markdown('<h5>NA</h5>', unsafe_allow_html=True)
+                    get_system_stat(stframe1, stframe2, stframe3, fps, df_fq)
